sửa sql

+++ DTO/DTO_Kho.cs
@@ -10,7 +10,7 @@ namespace DTO
     {
         public string MaNL { get; set; }
         public string TenNL { get; set; }
-        public int SoLuongTon { get; set; }
+        public decimal SoLuongTon { get; set; }
         public string DonViTinh { get; set; }
         public DateTime NgayNhap { get; set; }
         public string TenNCC { get; set; }
